before chat

+++ src/main/java/com/athul/memegramspring/service/impl/CommentServiceImpl.java
@@ -6,6 +6,7 @@ import com.athul.memegramspring.dto.UserDTO;
 import com.athul.memegramspring.entity.Comment;
 import com.athul.memegramspring.entity.Post;
 import com.athul.memegramspring.entity.User;
+import com.athul.memegramspring.exceptions.PermissionDeniedException;
 import com.athul.memegramspring.exceptions.ResourceNotFoundException;
 import com.athul.memegramspring.repository.CommentRepo;
 import com.athul.memegramspring.repository.PostRepo;
@@ -17,6 +18,8 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -29,7 +32,6 @@ public class CommentServiceImpl implements CommentService {
 
     @Override
     public CommentDTO addComment(int postId, int userId, String commentText) {
-
 
         String errorCode = "CommentServiceImpl:addComment()";
 
@@ -45,16 +47,28 @@ public class CommentServiceImpl implements CommentService {
 
         ModelMapper modelMapper = new ModelMapper();
         CommentDTO addedCommentDto = modelMapper.map(addedComment, CommentDTO.class);
+        addedCommentDto.setCommentedDate(addedComment.getCommentedDate());
         addedCommentDto.setUser(modelMapper.map(commentedUser, UserDTO.class));
         addedCommentDto.setPost(modelMapper.map(commentedPost, PostDTO.class));
         return addedCommentDto;
     }
 
     @Override
-    public void deleteComment(int commentId) {
+    public void deleteComment(int commentId, String loggedInUserName) {
         String errorCode = "CommentServiceImpl:deleteComment()";
         Comment comment = commentRepo.findById(commentId).orElseThrow(()->new ResourceNotFoundException("Comment","Id",commentId,errorCode));
-        commentRepo.delete(comment);
+        if(!comment.getUser().getEmail().equals(loggedInUserName)){
+            throw new PermissionDeniedException("User doesn't have permission to delete this comment");
+        }else {
+            commentRepo.delete(comment);
+        }
+    }
 
+    @Override
+    public List<CommentDTO> getAllComments(int postId) {
+        List<Comment> comment = commentRepo.findByPost(postId);
+        ModelMapper modelMapper = new ModelMapper();
+        List<CommentDTO> commentDTOS = comment.stream().map(x->modelMapper.map(x, CommentDTO.class)).collect(Collectors.toList());
+        return commentDTOS;
     }
 }
